Reuse callbacks_test, siplify memore mocking setup
