+++ test/test_allocator.h
@@ -26,11 +26,13 @@ void *instrumented_malloc(size_t size);
 
 void *instrumented_realloc(void *ptr, size_t size);
 
-#define WITH_MOCK_MALLOC(block, malloc_calls, ...) \
-  do {                                             \
-    set_mock_malloc(malloc_calls, __VA_ARGS__);    \
-    block;                                         \
-    finalize_mock_malloc();                        \
+#define WITH_MOCK_MALLOC(block, malloc_calls, ...)                    \
+  do {                                                                \
+    cbor_set_allocs(instrumented_malloc, instrumented_realloc, free); \
+    set_mock_malloc(malloc_calls, __VA_ARGS__);                       \
+    block;                                                            \
+    finalize_mock_malloc();                                           \
+    cbor_set_allocs(malloc, realloc, free);                           \
   } while (0)
 
 #define WITH_FAILING_MALLOC(block) WITH_MOCK_MALLOC(block, 1, MALLOC_FAIL)
